+++ src/main/java/org/example/elearningbe/course/dto/CourseResponse.java
@@ -1,12 +1,14 @@
 package org.example.elearningbe.course.dto;
 
+import java.math.BigDecimal;
+
 public record CourseResponse(
         Long id,
         String title,
         String description,
         String image,
         String category,
-        float price,
+        BigDecimal price,
         Integer duration,
         String ownerEmail
 ){
