+++ src/main/java/org/example/elearningbe/payment/cart/dto/CartItemResponse.java
@@ -1,9 +1,11 @@
 package org.example.elearningbe.payment.cart.dto;
 
+import java.math.BigDecimal;
+
 public record CartItemResponse(
         Long id,
         Long courseId,
         String courseTitle,
-        float price,
+        BigDecimal price,
         int quantity
 ) {}
