+++ src/main/java/org/example/elearningbe/payment/cart/dto/CartResponse.java
@@ -1,9 +1,10 @@
 package org.example.elearningbe.payment.cart.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CartResponse(
         Long id,
         List<CartItemResponse> items,
-        float totalAmount
+        BigDecimal totalAmount
 ) {}
