+++ src/main/java/org/example/elearningbe/payment/order/dto/OrderItemResponse.java
@@ -1,8 +1,10 @@
 package org.example.elearningbe.payment.order.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItemResponse(
         Long id,
         Long courseId,
         String courseTitle,
-        float priceAtPurchase
+        BigDecimal priceAtPurchase
 ) {}
