+++ src/main/java/org/example/elearningbe/payment/order/dto/OrderResponse.java
@@ -1,10 +1,11 @@
 package org.example.elearningbe.payment.order.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderResponse(
         Long id,
-        float totalAmount,
+        BigDecimal totalAmount,
         String status,
         List<OrderItemResponse> items
 ) {}
